Split semantic types following task-list/semantic structure
