+++ src/langium-model-server/semantic/semantic-model-index.ts
@@ -1,0 +1,10 @@
+import type { NamedSemanticElement } from './semantic-model'
+
+export type SemanticIndex = {
+    readonly id: string
+    findElementByName(name: string): NamedSemanticElement | undefined
+}
+
+export type ModelAwareSemanticIndex<SemI extends SemanticIndex> = SemI & {
+    readonly model: object
+}
