+++ src/langium-model-server/semantic/semantic-model.ts
@@ -1,0 +1,5 @@
+
+export type NamedSemanticElement = {
+    readonly id: string
+    name: string
+}
